change to FFTW_ESTIMATE
added a macro in tIDLib.h called FFTWPLANNERFLAG and used it in all sources so that either MEASURE or ESTIMATE can be chosen in the header. and going with ESTIMATE for now because the measurement process takes quite a while for very large FFTs. Not sure that the benefits of measurement are worth that. Ideally, this would be something that could be set per object. Not up for doing that right now.